chore(migration.sql): cast the count result to integer in the "admins" column to ensure correct data type

+++ src/server/models/db/migrations/20230829153602_add_users_to_lists_for_dashboard_view/migration.sql
@@ -23,7 +23,7 @@ create or replace view "ListsForDashboard" as (
     coalesce(li.live, 0) as live,
     coalesce(li."actionNeeded", 0) as "actionNeeded",
     coalesce(l."lastAnnualReviewStartDate", fp."firstPublished", current_date) + interval '18 months' < current_date as "isOverdue",
-    coalesce(count((select array_agg(u.email) from "_ListToUser" lu inner join "User" u on lu."B" = u.id where lu."A" = l.id)), 0) as "admins", -- updated column
+    coalesce(count((select array_agg(u.email) from "_ListToUser" lu inner join "User" u on lu."B" = u.id where lu."A" = l.id)), 0)::integer as "admins", -- updated column
     coalesce((select array_agg(u.email) from "_ListToUser" lu inner join "User" u on lu."B" = u.id where lu."A" = l.id), '{}'::varchar[]) as "userEmails" -- new column
   from
     "List" l
